+++ drivers/kafka/driver.mk
@@ -4,7 +4,7 @@ PROBE.kafka = docker exec kafkaJson kafka-topics --bootstrap-server localhost:90
 
 # Kafka-only: consumer-group rebalance recovery (CI: kafka-rebalance-tests.yml).
 .PHONY: test.kafka.rebalance
-test.kafka.rebalance: db.kafka.start db.destination.all.start $(ICEBERG_JAR)
+test.kafka.rebalance: olake.kafka.start olake.destination.all.start $(ICEBERG_JAR)
 	cd tests && go test -v ./kafka/... -timeout 0 -count=1 -run 'Rebalance'
 
 HELP_TARGETS += test.kafka.rebalance
