+++ drivers/mssql/internal/config_test.go
@@ -5,8 +5,8 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/datazip-inc/olake/lib"
 	"github.com/datazip-inc/olake/utils"
+	"github.com/datazip-inc/olake/utils/testutils"
 	"github.com/stretchr/testify/require"
 )
 
@@ -257,7 +257,7 @@ func TestConfig_URI(t *testing.T) {
 		{
 			name: "with jdbc url params",
 			config: func() *Config {
-				certs := lib.GenerateTestCerts()
+				certs := testutils.GenerateTestCerts()
 				return &Config{
 					Host:     "mssql-host",
 					Port:     1433,
