+++ drivers/platforms.conf
@@ -1,5 +1,5 @@
 # Docker build platforms per driver ('*' = default).
-# Used by release-tool.sh and `make docker.build.<driver>`.
+# Used by release-tool.sh and `make docker.<driver>.build`.
 * = linux/amd64,linux/arm64
 # IBM provides no ARM64 clidriver.
 db2 = linux/amd64
