+++ lib/go.mod
@@ -1,5 +1,3 @@
 module github.com/datazip-inc/olake/lib
 
 go 1.25.12
-
-require golang.org/x/sync v0.21.0
