+++ tests/mysql/mysql_util_test.go
@@ -7,7 +7,6 @@ import (
 	"time"
 
 	"github.com/apache/arrow-go/v18/arrow"
-	"github.com/datazip-inc/olake/lib/utils"
 	"github.com/datazip-inc/olake/tests/testutils"
 	_ "github.com/go-sql-driver/mysql"
 	"github.com/jmoiron/sqlx"
@@ -21,7 +20,7 @@ func ExecuteQuery(ctx context.Context, t *testing.T, streams []string, operation
 	var connStr string
 	if fileConfig {
 		var config Config
-		utils.UnmarshalFile("./testdata/source.json", &config, false)
+		testutils.UnmarshalFile("./testdata/source.json", &config, false)
 		connStr = fmt.Sprintf("%s:%s@tcp(%s:%d)/%s?parseTime=true",
 			config.Username,
 			config.Password,
@@ -250,7 +249,7 @@ func ExecuteQuery(ctx context.Context, t *testing.T, streams []string, operation
 	case "bulk_cdc_data_insert":
 		backfillStreams := testutils.GetBackfillStreamsFromCDC(streams)
 		// insert the data into the cdc tables concurrently
-		err := utils.Concurrent(ctx, streams, len(streams), func(ctx context.Context, cdcStream string, executionNumber int) error {
+		err := testutils.Concurrent(ctx, streams, len(streams), func(ctx context.Context, cdcStream string, executionNumber int) error {
 			_, err = db.ExecContext(ctx, fmt.Sprintf("INSERT INTO %s SELECT * FROM %s LIMIT 15000000", cdcStream, backfillStreams[executionNumber]))
 			if err != nil {
 				return err
