+++ utils/utils.go
@@ -17,7 +17,6 @@ import (
 	"time"
 
 	"github.com/datazip-inc/olake/constants"
-	libutils "github.com/datazip-inc/olake/lib/utils"
 	"github.com/datazip-inc/olake/utils/logger"
 	"github.com/goccy/go-json"
 	"github.com/oklog/ulid"
@@ -68,12 +67,23 @@ func ArrayContains[T any](set []T, match func(elem T) bool) (int, bool) {
 
 // returns cond ? a ; b (note: it is not function ternary)
 func Ternary(cond bool, a, b any) any {
-	return libutils.Ternary(cond, a, b)
+	if cond {
+		return a
+	}
+	return b
 }
 
 // return the average of the given values.
 func Average[T int | int8 | int16 | int32 | int64 | float32 | float64](values []T) float64 {
-	return libutils.Average(values)
+	if len(values) == 0 {
+		return 0.0
+	}
+
+	var sum float64
+	for _, v := range values {
+		sum += float64(v)
+	}
+	return sum / float64(len(values))
 }
 
 func ForEach[T any](set []T, action func(elem T) error) error {
@@ -332,12 +342,52 @@ func ComputeConfigHash(srcPath, destPath string) string {
 }
 
 func NormalizedEqual(strune1, strune2 string) bool {
-	return libutils.NormalizedEqual(strune1, strune2)
+	normalize := func(s string) (string, error) {
+		// Slice out exactly from the first '{' to the last '}'
+		start := strings.IndexRune(s, '{')
+		end := strings.LastIndex(s, "}")
+		if start < 0 || end < 0 || start > end {
+			return "", fmt.Errorf("no valid JSON object found")
+		}
+		core := s[start : end+1]
+		// remove whitespace
+		core = strings.ReplaceAll(core, " ", "")
+		core = strings.ReplaceAll(core, "\n", "")
+		core = strings.ReplaceAll(core, "\t", "")
+		return core, nil
+	}
+
+	c1, err := normalize(strune1)
+	if err != nil {
+		return false
+	}
+	c2, err := normalize(strune2)
+	if err != nil {
+		return false
+	}
+
+	rune1 := []rune(c1)
+	rune2 := []rune(c2)
+	if len(rune1) != len(rune2) {
+		return false
+	}
+	sort.Slice(rune1, func(i, j int) bool { return rune1[i] < rune1[j] })
+	sort.Slice(rune2, func(i, j int) bool { return rune2[i] < rune2[j] })
+	return string(rune1) == string(rune2)
 }
 
 // Reformat makes all keys to lower case and replaces all special symbols with '_'
 func Reformat(key string) string {
-	return libutils.Reformat(key)
+	key = strings.ToLower(key)
+	var result strings.Builder
+	for _, symbol := range key {
+		if IsLetterOrNumber(symbol) {
+			result.WriteByte(byte(symbol))
+		} else {
+			result.WriteRune('_')
+		}
+	}
+	return result.String()
 }
 
 // IsLetterOrNumber returns true if input symbol is:
